requirements.txt
neo4j>=5.22.0
neo4j-graphrag>=0.4.0
python-dotenv>=1.0.1
graphdatascience>=1.14,<2.0
